# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CutAI - AI Video Editor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exports:
    post:
      operationId: createExport
      tags: [exports]
      summary: Start an export job for a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExportBody"
      responses:
        "201":
          description: Created export
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Export"
  /exports/{id}:
    get:
      operationId: getExport
      tags: [exports]
      summary: Get export status and download URL
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Export details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Export"
        "404":
          description: Not found
  /projects/{id}/exports:
    get:
      operationId: listProjectExports
      tags: [exports]
      summary: List all exports for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of exports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Export"
...
components:
  schemas:
...
    Export:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        format:
          type: string
          enum: [mp4, mov, webm]
        resolution:
          type: string
          enum: [1080p, 720p, 4k, vertical_1080, vertical_720]
        status:
          type: string
          enum: [pending, rendering, completed, failed]
        progress:
          type: integer
        filePath:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        downloadUrl:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
      required:
        - id
        - projectId
        - format
        - resolution
        - status
        - progress
        - createdAt
    CreateExportBody:
      type: object
      properties:
        projectId:
          type: string
        format:
          type: string
          enum: [mp4, mov, webm]
        resolution:
          type: string
          enum: [1080p, 720p, 4k, vertical_1080, vertical_720]
      required:
        - projectId
        - format
        - resolution
...